Moved CRC32 implementation into its own "library", separated definitions into their own source files (utils.c, crc32.c)
